requirements.txt
pyyaml
python-magic